+++ src/main/java/jp/co/ricoh/cotos/commonlib/dto/json/estimation/ProductEstimationExtendsParameterO365Dto.java
@@ -1,5 +1,7 @@
 package jp.co.ricoh.cotos.commonlib.dto.json.estimation;
 
+import java.util.List;
+
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
@@ -22,12 +24,12 @@ public class ProductEstimationExtendsParameterO365Dto {
 	 * CPQ商品固有戻り値
 	 */
 	@JsonProperty("cpqProductSpecificReturnValue")
-	private CpqReturnValueO365Dto cpqReturnValueO365Dto;
+	private CpqReturnValueO365Dto cpqProductSpecificReturnValueDto;
 
 	/**
-	 * 移行用項目
+	 * 品種(見積用)
 	 */
-	@JsonProperty("migrationParameter")
-	private ProductEstimationMigrationParameterDto productEstimationMigrationParameterDto;
+	@JsonProperty("itemEstimation")
+	private List<ItemEstimationO365Dto> itemEstimationDtoList;
 
 }