+++ src/main/java/jp/co/ricoh/cotos/commonlib/entity/master/ProductMaster.java
@@ -302,8 +302,15 @@ public class ProductMaster extends EntityBaseMaster {
 	private SerialContactDiv serialContactDiv;
 
 	/**
+	 * 正式商品名
+	 */
+	@Column(nullable = false)
+	@ApiModelProperty(value = "正式商品名", required = true, position = 30, allowableValues = "range[0,255]")
+	private String officialProductName;
+
+	/**
 	 * RtoR設定可能区分
 	 */
-	@ApiModelProperty(value = "RtoR設定可能区分", required = false, position = 30, allowableValues = "引継ぎ先と引継ぎ元(\"1\"), 引継ぎ元のみ(\"2\")")
+	@ApiModelProperty(value = "RtoR設定可能区分", required = false, position = 31, allowableValues = "引継ぎ先と引継ぎ元(\"1\"), 引継ぎ元のみ(\"2\")")
 	private RtorSettingPossibleDiv rtorSettingPossibleDiv;
 }